// packages/dashboard/src/http/types/ListCompanies.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

export type ListCompaniesPath = {
    clientId: string;
};

export type ListCompaniesStatus200 = {
    data: {
        id: string;
        cnpj: string;
        razaoSocial: string;
        nomeFantasia: string | null;
        phone: string | null;
        email: string | null;
        address: string | null;
        isActive: boolean;
        legalRepresentativeId: string;
        /**
         * @description
         * Format: `date-time`
         * @type string
        */
        createdAt: string;
    }[];
};

export type ListCompaniesOptions = {
    body?: never;
    path: ListCompaniesPath;
    query?: never;
    headers?: never;
};

export type ListCompaniesResponses = {
    "200": ListCompaniesStatus200;
};

/**
 * @description Union of all possible responses
*/
export type ListCompaniesResponse = ListCompaniesStatus200;